Refactor multiple GCSE flows
- remove any dead code
