+++ app/controllers/candidate_interface/gcse/maths/grade_controller.rb
@@ -6,23 +6,23 @@ module CandidateInterface
     before_action :set_subject
 
     def edit
-      @application_qualification = details_form
-      @qualification_type = details_form.qualification.qualification_type
+      @application_qualification = maths_gsce_grade_form
+      @qualification_type = maths_gsce_grade_form.qualification.qualification_type
     end
 
     def update
-      @qualification_type = details_form.qualification.qualification_type
+      @qualification_type = maths_gsce_grade_form.qualification.qualification_type
 
-      details_form.grade = details_params[:grade]
-      details_form.other_grade = details_params[:other_grade]
+      maths_gsce_grade_form.grade = details_params[:grade]
+      maths_gsce_grade_form.other_grade = details_params[:other_grade]
 
-      @application_qualification = details_form.save_grade
+      @application_qualification = maths_gsce_grade_form.save_grade
 
       if @application_qualification
         update_gcse_completed(false)
         redirect_to next_gcse_path
       else
-        @application_qualification = details_form
+        @application_qualification = maths_gsce_grade_form
         track_validation_error(@application_qualification)
 
         render :edit
@@ -33,6 +33,24 @@ module CandidateInterface
 
     def set_subject
       @subject = 'maths'
+    end
+
+    def next_gcse_path
+      if maths_gsce_grade_form.award_year.nil?
+        candidate_interface_gcse_details_edit_year_path(subject: @subject)
+      else
+        candidate_interface_gcse_review_path(subject: @subject)
+      end
+    end
+
+    def maths_details_params
+      params.require(:candidate_interface_gcse_qualification_details_form).permit(%i[grade award_year other_grade])
+    end
+
+    def maths_gsce_grade_form
+      @maths_gcse_grade_form ||= GcseQualificationDetailsForm.build_from_qualification(
+        current_application.qualification_in_subject(:gcse, @subject),
+      )
     end
   end
 end
